+++ src/include/perft.h
@@ -4,7 +4,7 @@
 #include "chess.h"
 #include <inttypes.h>
 
-uint64_t perft(Position* position, const uint8_t depth);
-void     divide(Position* position, const uint8_t depth);
+unsigned long long perft(Position* position, const uint8_t depth);
+void               divide(Position* position, const uint8_t depth);
 
 #endif
